requirements.txt
cycler==0.10.0
decorator==4.4.1
imageio==2.8.0
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
opencv-python==4.2.0.32
pathlib==1.0.1
Pillow==7.0.0
PyOpenGL==3.1.5
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
trimesh==3.5.23
xxhash==1.4.3
